+++ lib/yield_star_client/extract_lease_term_rent_hashes.rb
@@ -10,12 +10,20 @@ module YieldStarClient
         to_hash[soap_wrapper_element][:return]
 
       external_property_id = return_element[:external_property_id]
-      lease_term_rent = return_element[soap_unit_element] || {}
 
-      unit_rates = [lease_term_rent[:unit_rate]].flatten.compact
+      ltr_unit_response_hashes = return_element[soap_unit_element] || {}
+      ltr_unit_response_hashes = [ltr_unit_response_hashes].flatten
+
+      ltr_unit_response_hashes.inject([]) do |lease_term_rent_hashes, ltr_unit_response_hash|
+        lease_term_rent_hashes + convert_response_hash_into_ltr_hashes(ltr_unit_response_hash, external_property_id)
+      end
+    end
+
+    def self.convert_response_hash_into_ltr_hashes(ltr_unit_response_hash, external_property_id)
+      unit_rates = [ltr_unit_response_hash[:unit_rate]].flatten.compact
       unit_rates.map do |unit_rate|
         hash = unit_rate.
-          merge(lease_term_rent.reject { |k, v| k == :unit_rate }).
+          merge(ltr_unit_response_hash.reject { |k, v| k == :unit_rate }).
           merge(external_property_id: external_property_id)
       end
     end