+++ lib/yield_star_client/lease_term_request_config.rb
@@ -12,6 +12,7 @@ module YieldStarClient
         class_attribute :lease_term_request_opts
 
         attribute :external_property_id, String
+        attribute :units
         attribute :unit_number, String
         attribute :building, String
         attribute :min_lease_term, String
@@ -19,22 +20,35 @@ module YieldStarClient
         attribute :first_move_in_date, Date
         attribute :last_move_in_date, Date
 
-        validates :external_property_id, :unit_number, presence: true
+        validates :external_property_id, presence: true
       end
 
       def request_args
-        lease_term_rent_options = LeaseTermRentOptions.new(attributes.slice(
-          :unit_number,
-          :building,
-          :min_lease_term,
-          :max_lease_term,
-          :first_move_in_date,
-          :last_move_in_date,
-        ))
+        if units
+          lease_term_rent_options = [units].flatten.map do |options_hash|
+            options_hash.slice(
+              :unit_number,
+              :building,
+              :min_lease_term,
+              :max_lease_term,
+              :first_move_in_date,
+              :last_move_in_date,
+            )
+          end
+        else
+          lease_term_rent_options = LeaseTermRentOptions.new(attributes.slice(
+            :unit_number,
+            :building,
+            :min_lease_term,
+            :max_lease_term,
+            :first_move_in_date,
+            :last_move_in_date,
+          )).to_request_hash
+        end
         request_element = self.class.lease_term_request_opts.
           fetch(:request_element)
         attributes.merge(
-           request_element => lease_term_rent_options.to_request_hash
+           request_element => lease_term_rent_options
         )
       end
     end