+++ spec/yield_star_client/lease_term_rent_methods_spec.rb
@@ -10,12 +10,17 @@ module YieldStarClient
       ))
     end
 
-    describe "#get_lease_term_rent" do
-      it "returns the lease term rents", vcr: {record: :once} do
-        properties = client.get_properties
-        external_property_id = properties.last.external_property_id
-        floor_plan = client.get_available_units(external_property_id).first
-        unit_number = floor_plan.units.first.unit_number
+    # NOTE: ensure that the units you use have lease term rents set up
+    describe "#get_lease_term_rent", vcr: {record: :once} do
+      let(:properties) { client.get_properties }
+      let(:external_property_id) { properties.last.external_property_id }
+      let(:floor_plan) { client.get_available_units(external_property_id).last }
+      let(:unit_number) { unit.unit_number }
+      let(:unit) { floor_plan.units.first }
+      let(:unit_number_2) { unit_2.unit_number }
+      let(:unit_2) { floor_plan.units.last }
+
+      it "returns the lease term rents" do
         lease_term_rents = client.get_lease_term_rent(
           external_property_id,
           unit_number,
@@ -29,6 +34,72 @@ module YieldStarClient
         let(:request_class) { GetLeaseTermRent::Request }
         let(:response_class) { GetLeaseTermRent::Response }
         let(:response_accessor) { :lease_term_rents }
+      end
+
+      context "when receiving with a hash parameter" do
+        let(:date_tomorrow) { Date.today + 1 }
+        it "returns lease terms for a single unit" do
+          lease_term_rents = client.get_lease_term_rent(
+            external_property_id: external_property_id,
+            units: {
+              # building: 1, # optional, and not needed in current vcr record
+              unit_number: unit_number_2,
+              min_lease_term: 1,
+              max_lease_term: 18,
+              first_move_in_date: date_tomorrow.to_s,
+              last_move_in_date: (date_tomorrow + 45).to_s,
+              ready_for_move_in_date: date_tomorrow,
+              unit_available_date: date_tomorrow,
+            }
+          )
+
+          expect(lease_term_rents).to_not be_empty
+          expect(lease_term_rents.first).to be_a LeaseTermRent
+          lease_term_rents.each do |lt|
+            expect(lt.unit_number).to eq unit_number_2
+            expect(lt.lease_term).to be >= 1
+            expect(lt.lease_term).to be <= 18
+            expect(lt.move_in_date).to be >= date_tomorrow
+            expect(lt.move_in_date).to be <= (date_tomorrow + 45)
+          end
+        end
+
+        it "returns lease terms for multiple units" do
+          lease_term_rents = client.get_lease_term_rent(
+            external_property_id: external_property_id,
+            units: [
+              {
+                # building: 1, # optional, and not needed in current vcr record
+                unit_number: unit_number,
+                min_lease_term: 1,
+                max_lease_term: 18,
+                first_move_in_date: date_tomorrow.to_s,
+                last_move_in_date: (date_tomorrow + 45).to_s,
+                ready_for_move_in_date: date_tomorrow,
+                unit_available_date: date_tomorrow,
+              },
+              {
+                unit_number: unit_number_2,
+                min_lease_term: 1,
+                max_lease_term: 18,
+                first_move_in_date: date_tomorrow.to_s,
+                last_move_in_date: (date_tomorrow + 45).to_s,
+                ready_for_move_in_date: date_tomorrow,
+                unit_available_date: date_tomorrow,
+              }
+            ]
+          )
+
+          expect(lease_term_rents).to_not be_empty
+          expect(lease_term_rents.first).to be_a LeaseTermRent
+          lease_term_rents.each do |lt|
+            expect([unit_number, unit_number_2]).to include lt.unit_number
+            expect(lt.lease_term).to be >= 1
+            expect(lt.lease_term).to be <= 18
+            expect(lt.move_in_date).to be >= date_tomorrow
+            expect(lt.move_in_date).to be <= (date_tomorrow + 45)
+          end
+        end
       end
     end